feat: Atualiza endpoint Administrator::Biddings::RegenerateMinutesController para considerar as situações Fracassada, Deserta e Concluída
